@@ -11,7 +11,7 @@
     "build": "tsc",
     "docs": "typedoc src",
     "lint": "eslint .",
-    "test": "c8 node --unhandled-rejections=strict test"
+    "test": "c8 tsx test"
   },
   "keywords": [
     "static",
@@ -34,21 +34,23 @@
   },
   "license": "MIT",
   "dependencies": {
-    "mime": "^3.0.0",
-    "minimist": "^1.2.7"
+    "mime": "^4.0.4",
+    "minimist": "^1.2.8"
   },
   "devDependencies": {
-    "@types/mime": "^3.0.1",
-    "@types/node": "^18.11.18",
-    "@typescript-eslint/eslint-plugin": "^5.49.0",
-    "@typescript-eslint/parser": "^5.49.0",
-    "axios": "^1.2.4",
-    "basictap": "^3.4.4",
-    "c8": "^7.12.0",
+    "@types/mime": "^4.0.0",
+    "@types/node": "^22.5.4",
+    "@typescript-eslint/eslint-plugin": "^8.5.0",
+    "@typescript-eslint/parser": "^8.5.0",
+    "axios": "^1.7.7",
+    "basictap": "^4.3.5",
+    "c8": "^10.1.2",
+    "eslint": "^8.57.0",
     "fetch-h2": "^3.0.2",
-    "semistandard": "^16.0.1",
-    "typedoc": "^0.23.24",
-    "typescript": "^4.9.4"
+    "semistandard": "^17.0.0",
+    "tsx": "^4.19.1",
+    "typedoc": "^0.26.7",
+    "typescript": "^5.6.2"
   },
   "directories": {
     "test": "test"
